Added NarrowPhase::CollideShapeWithInternalEdgeRemoval
This function allows you to do a collision test that will automatically remove internal edges from the collision results.
